% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_soil_mod.R
\name{append_soil_mod}
\alias{append_soil_mod}
\title{Append Soil Mod}
\usage{
append_soil_mod(
  crop_data = croptimizer::crops,
  soil_mods = NULL,
  soil_mod_names = croptimizer::soil_mod_names
)
}
\arguments{
\item{crop_data}{a data.frame containing the data about the crops,
expected to contain the same columns as `croptimizer::crops`}
}
\value{
crop_data with a soil_mod column appended based upon user
input if provided, and containing "normal" if not
}
\description{
Appends a column for soil modification (such as fertilizer or speed-gro)
to the crop data for later use in calculating what fits in season (using
`filter_to_season()` and `filter_to_days()`) and in calculating what the
expected sale price is (using `calc_crop_sale_price()`).
}
