% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_crop_sale_price.R
\name{calc_crop_sale_price}
\alias{calc_crop_sale_price}
\title{Calculate Crop Sale Price}
\usage{
calc_crop_sale_price(
  crop_data = croptimizer::crops,
  farming_level = 0,
  level_5_tiller = NULL
)
}
\arguments{
\item{crop_data}{data.frame containing crop data, which must contain the
columns `sell_price`, `chance_for_extra_crops`, `max_extra_harvest`}

\item{farming_level}{integer representing current farming level}

\item{level_5_tiller}{logical representing taking the tiller specialization
at level 5}
}
\value{

}
\description{
Calculates the expected value of a crop to sell, taking into account
the probability of extra harvests and the number of crops in an extra
harvest.
}
\examples{
## Calculate crop sale prices if you have tiller specialty
calc_crop_sale_price(farming_level = 5, level_5_tiller = TRUE)

## Calculate crop sale price if you did not take tiller
calc_crop_sale_price(farming_level = 10, level_5_tiller = FALSE)
}
