% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_net_profit.R
\name{calc_net_profit}
\alias{calc_net_profit}
\title{Calculate Net Profit}
\usage{
calc_net_profit(crop_data, sort = TRUE)
}
\arguments{
\item{crop_data}{crop data to sort, must contain columns `seed_price` and
`expected_sell_price`}

\item{sort}{a logical indicating if you would like the results sorted by net
profit}
}
\value{
crop_data with a net profit column appended
}
\description{
Calculates the net profit for a crop by subtracting the cost of
the seeds from the sale price of the crop. Does not currently
account for the cost of soil modifiers such as fertilizer and
speed-gro, and does not account for seeds potentially obtained
for free.
}
\details{
Crop data is required to contain columns `seed_price` and
`expected_sell_price`.
}
