% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croptimize.R
\name{croptimize}
\alias{croptimize}
\title{Croptmize}
\usage{
croptimize(
  joja_member = FALSE,
  soil_mods = NULL,
  season,
  day_of_season,
  farming_level = 0,
  level_5_tiller = NULL,
  crops = croptimizer::crops
)
}
\value{

}
\description{
Higher-level function for sorting available crops by most profitable. Most users will want to use
this function.
}
