% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_to_days.R
\name{filter_to_days}
\alias{filter_to_days}
\title{Filter To Days}
\usage{
filter_to_days(day_of_season, crop_data = croptimizer::crops)
}
\arguments{
\item{day_of_season}{integer or numeric representing the date (current day
number) of current season}

\item{crop_data}{data.frame containing crop data, defaults to package data
`crops`}
}
\value{
data.frame containing crop data filtered to crops which have
sufficient time left in the season to grow
}
\description{
Filters crop dataset to only crops which have sufficient time left to grow.
}
\examples{
filter_to_days(day = 25)
}
