% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_to_season.R
\name{filter_to_season}
\alias{filter_to_season}
\title{Filter To Season}
\usage{
filter_to_season(season, crop_data = croptimizer::crops)
}
\arguments{
\item{season}{character containing current season}

\item{crop_data}{data.frame containing crop data, defaults to package data `crops`}
}
\value{
data.frame containing crop data filtered to current season crops
}
\description{
Filters crop data to only crops which grow in the specified season.
}
\examples{
## Filters to fall season
filter_to_season(season = "fall")
}
