% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_total_days_in_growth.R
\name{update_total_days_in_growth}
\alias{update_total_days_in_growth}
\title{Update Total Days in Growth}
\usage{
update_total_days_in_growth(crop_data, level_10_agri = NULL)
}
\value{
crop data with updated `total_days_in_growth` which accounts for
Speed-Gro
}
\description{
Updates the `total_days_in_growth` column in crops data to account
for decreased days in growth due to applying Speed-Gro. Crop data
must contain the column `soil_mod`, indicating what fertilizer of speed-gro
is applied to that crop.
}
\examples{
## Hyper Speed-Gro
speed_gro <-
append_soil_mod(soil_mods = "hyper speed-gro") \%>\%
update_total_days_in_growth()
## Deluxe Speed-Gro
speed_gro <-
append_soil_mod(soil_mods = "deluxe speed-gro") \%>\%
update_total_days_in_growth()
}
