% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_crop_sale_price.R
\name{calc_crop_sale_price}
\alias{calc_crop_sale_price}
\title{Calculate Crop Sale Price}
\usage{
calc_crop_sale_price(crop_data = croptimizer::crops)
}
\arguments{
\item{crop_data}{data.frame containing crop data, which must contain the
columns `sell_price`, `chance_for_extra_crops`, `max_extra_harvest`}
}
\value{

}
\description{
Calculates the expected value of a crop to sell, taking into account
the probability of extra harvests and the number of crops in an extra
harvest.
}
