% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_seed_prices.R
\name{calc_seed_prices}
\alias{calc_seed_prices}
\title{Calc Seed Prices}
\usage{
calc_seed_prices(
  crop_data = croptimizer::crops,
  joja_member = FALSE,
  drop_cols = FALSE
)
}
\arguments{
\item{crop_data}{data.frame containing crop data, defaults to the data
contained in the croptimizer package}

\item{joja_member}{logical specifying if you have the JojaMart membership}
}
\value{
data.frame containing crop_data with seed_price and source columns
appended
}
\description{
Calc Seed Prices
}
\examples{
fall_crops <-
filter_to_season(season = "fall", crop_data = croptimizer::crops)
calc_seed_prices(crop_data = fall_crops,
joja_member = FALSE,
drop_cols = TRUE)
}
